requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-crispy-forms==1.7.2
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
Markdown==2.6.11
oauthlib==2.1.0
Pillow==5.2.0
psycopg2==2.7.5
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23
